requirements.txt
flask~=2.0.1
Flask-MongoAlchemy
pymongo~=2.8.1
PyJWT~=2.1.0
Werkzeug~=2.0.1